agentic-move-agent/requirements.txt
gradio==4.44.0
boto3==1.35.0
pillow==10.4.0
python-dotenv==1.0.0
anthropic==0.39.0
